Organize data loader and data mapping
- Moves mappings into their own folder
- Clean up, removes unused resources
- Reorganize data loader to make it easier to follow
- Breaking it out into files for readability
- Adds tests
- Account for some edge cases
