+++ backend/config/settings.py
@@ -84,14 +84,7 @@ INSTALLED_APPS += [
 ]
 
 # Our apps
-INSTALLED_APPS += [
-    "audit",
-    "api",
-    "users",
-    "report_submission",
-    "cms",
-    "data_distro"
-]
+INSTALLED_APPS += ["audit", "api", "users", "report_submission", "cms", "data_distro"]
 
 MIDDLEWARE = [
     "django.middleware.security.SecurityMiddleware",
@@ -245,7 +238,10 @@ else:
     SESSION_COOKIE_SAMESITE = "Lax"
     X_FRAME_OPTIONS = "DENY"
 
-    CORS_ALLOWED_ORIGINS = [f"https://{AWS_S3_CUSTOM_DOMAIN}", env.str("DJANGO_BASE_URL")]
+    CORS_ALLOWED_ORIGINS = [
+        f"https://{AWS_S3_CUSTOM_DOMAIN}",
+        env.str("DJANGO_BASE_URL"),
+    ]
     CORS_ALLOW_METHODS = ["GET", "OPTIONS"]
 
 ADMIN_URL = "admin/"
