+++ backend/data_distro/data_to_load/test_data/agency.txt
@@ -1,0 +1,11 @@
+AUDITYEAR|DBKEY|EIN|AGENCY
+1997|100000|730776899|00
+1998|100000|730776899|04
+1999|100000|730776899|00
+2000|100000|730776899|
+2001|100000|730776899|00
+2002|100000|730776899|07
+2003|100000|730776899|00
+2004|100000|730776899|
+2005|100000|730776899|00
+2006|100000|730776899|10
