+++ backend/data_distro/data_to_load/test_data/cpas.txt
@@ -1,0 +1,3 @@
+DBKEY|AUDITYEAR|CPAFIRMNAME|CPAEIN|CPASTREET1|CPACITY|CPASTATE|CPAZIPCODE|CPACONTACT|CPATITLE|CPAPHONE|CPAFAX|CPAEMAIL
+100000|2012|JOHN FLUSCHE, CPA||6851 SOUTH CANTON AVENUE|TULSA|OK|741363405|JOHN FLUSCHE|CPA|9184888753|9184889256|CPA@OKLAHOMA.NET
+100000|2014|Paul Bunyan, CPA|431026561|6851 SOUTH CANTON|TULSA|OK|74136|JOHN FLUSCHE|CPA|9184888753|9184889256|CPA@OKLAHOMA.NET
