--- backend/data_distro/download_model_dictonaries.py
@@ -1,280 +1,0 @@
-# This is using the data download key info to map out the data
-
-
-# I plan on updating these as I improve models so we can use this for an import script
-model_transforms = [
-    {"GENERAL": "General"},
-    {"CFDA INFO": "CfdaInfo"},
-    {"FINDINGS": "Findings"},
-    {"FINDINGSTEXT": "Findingstext"},
-    {"CAPTEXT": "Captext"},
-    {"NOTES": "Notes"},
-    {"MULTIPLE CPAS INFO": "MultipleCpasInfo"},
-    {"REVISIONS": "Revisions"},
-    {"UEI INFO": "UeiInfo"},
-    {"AGENCIES": "Agencies"},
-    {"PASSTHROUGH": "Passthrough"},
-    {"EIN INFO": "EinInfo"},
-    {"DUN INFO": "DunsInfo"},
-]
-
-model_title_transforms = [
-    {"gen": "General"},
-    {"cfda": "CfdaInfo"},
-    {"findings": "Findings"},
-    {"findingstext_formatted": "Findingstext"},
-    {"captext_formatted": "Captext"},
-    {"notes": "Notes"},
-    {"cpas": "MultipleCpasInfo"},
-    {"revisions": "Revisions"},
-    {"ueis": "UeiInfo"},
-    {"agency": "Agencies"},
-    {"passthrough": "Passthrough"},
-    {"eins": "EinInfo"},
-    {"duns": "DunInfo"},
-]
-
-field_mappings = {
-    #
-    "AGENCY": "agency_cfda",
-    "AGENCYCFDA": "agency_cfda",
-    "AMOUNT": "amount",
-    "ARRA": "arra",
-    "AUDITEECERTIFYNAME": "auditee_certify_name",
-    "AUDITEECERTIFYTITLE": "auditee_certify_title",
-    "AUDITEECONTACT": "auditee_contact",
-    "AUDITEEDATESIGNED": "auditee_date_signed",
-    "AUDITEEEMAIL": "auditee_email",
-    "AUDITEEFAX": "auditee_fax",
-    "AUDITEENAME": "auditee_name",
-    "AUDITEENAMETITLE": "auditee_name_title",
-    "AUDITEEPHONE": "auditee_phone",
-    "AUDITEETITLE": "auditee_title",
-    "AUDITINFO": "audit_info",
-    "AUDITINFO_EXPLAIN": "auditinfo_explain",
-    "AUDITOR_EIN": "auditor_ein",
-    "AUDITTYPE": "audit_type",
-    "AUDITYEAR": "audit_year",
-    "AWARDIDENTIFICATION": "award_identification",
-    "CAP": "cap",
-    "CAP_EXPLAIN": "cap_explain",
-    "CFDA": "cfda",
-    "CFDAPROGRAMNAME": "cfda_program_name",
-    "CHARTSTABLES": "charts_tables",
-    "CITY": "city",
-    "CLUSTERNAME": "cluster_name",
-    "CLUSTERTOTAL": "cluster_total",
-    "COGAGENCY": "cognizant_agency",
-    # not in key check
-    "COG_OVER": "cognizant_agency_over",
-    "COMPLETED_ON": "completed_on",
-    "COMPONENT DATE RECEIVED": "component_date_received",
-    "CONTENT": "content",
-    "CPACITY": "cpa_city",
-    "CPACONTACT": "cpa_contact",
-    "CPACOUNTRY": "cpa_country",
-    "CPADATESIGNED": "cpa_date_signed",
-    "CPAEIN": "cpa_ein",
-    "CPAEMAIL": "cpa_email",
-    "CPAFAX": "cpa_fax",
-    "CPAFIRMNAME": "cpa_firm_name",
-    "CPAFOREIGN": "cpa_foreign",
-    "CPAPHONE": "cpa_phone",
-    "CPASTATE": "cpa_state",
-    "CPASTREET1": "cpa_street1",
-    "CPASTREET2": "cpa_street2",
-    "CPATITLE": "cpa_title",
-    "CPAZIPCODE": "cpa_zip_code",
-    "CYFINDINGS": "current_or_former_findings",
-    "DATEFIREWALL": "date_firewall",
-    "DBKEY": "dbkey",
-    "DIRECT": "direct",
-    "DOLLARTHRESHOLD": "dollar_threshold",
-    "DUNS": "duns",
-    "DUNSEQNUM": "duns_seq_num",
-    "DUP_REPORTS": "dup_reports",
-    "EIN": "ein",
-    "EINSEQNUM": "ein_seq_num",
-    "EINSUBCODE": "ein_subcode",
-    "ELECAUDITFINDINGSID": "elec_audit_findings_id",
-    "ELECAUDITSID": "elec_audits_id",
-    "ELECRPTREVISIONID": "elec_report_revision_id",
-    "ENTITY_TYPE": "entity_type",
-    # only in key
-    "FAC ACCEPTED DATE": "fac_accepted_date",
-    # not in key but not worried FAC ACCEPTED DATE was not
-    "FACACCEPTEDDATE": "fac_accepted_date",
-    "FEDERALAWARDS": "federal_awards",
-    "FEDERALAWARDS_EXPLAIN": "federal_awards_explain",
-    "FEDERALPROGRAMNAME": "federal_program_name",
-    "FINDINGREFNUMS": "finding_ref_nums",
-    # not in key check
-    "FINDINGSREFNUMS": "finding_ref_nums",
-    "FINDINGS": "findings",
-    "FINDINGSCOUNT": "findings_count",
-    "FINDINGSTEXT": "findings_text",
-    "FINDINGSTEXT_EXPLAIN": "findings_text_explain",
-    "FINDINGS_EXPLAIN": "findings_explain",
-    "FORM DATE RECEIVED": "form_date_received",
-    "FYENDDATE": "fy_end_date",
-    "FYSTARTDATE": "fy_start_date",
-    "GENINFO": "general_info",
-    "GENINFO_EXPLAIN": "general_info_explain",
-    "GOINGCONCERN": "going_concern",
-    "ID": "fac_id",
-    "INITIAL DATE RECEIVED": "initial_date_received",
-    "LOANBALANCE": "loan_balance",
-    "LOANS": "loans",
-    "LOWRISK": "low_risk",
-    "MAJORPROGRAM": "major_program",
-    "MATERIALNONCOMPLIANCE": "material_noncompliance",
-    "MATERIALWEAKNESS": "material_weakness",
-    "MATERIALWEAKNESS_MP": "material_weakness_major_program",
-    "MODIFIEDOPINION": "modified_opinion",
-    "MULTIPLEDUNS": "multiple_duns",
-    "MULTIPLEEINS": "multiple_eins",
-    "MULTIPLEUEIS": "multiple_ueis",
-    "MULTIPLE_CPAS": "multiple_cpas",
-    "NOTESTOSEFA": "notes_to_sefa",
-    "NOTESTOSEFA_EXPLAIN": "notes_to_sefa_explain",
-    "NOTE_INDEX": "note_index",
-    "NUMBERMONTHS": "number_months",
-    "OTHER": "other",
-    "OTHERCLUSTERNAME": "other_cluster_name",
-    "OTHERFINDINGS": "other_findings",
-    "OTHERNONCOMPLIANCE": "other_non_compliance",
-    "OTHER_EXPLAIN": "other_explain",
-    "OVERSIGHTAGENCY": "oversight_agency",
-    "PASSTHROUGHAMOUNT": "passthrough_amount",
-    "PASSTHROUGHAWARD": "passthrough_award",
-    "PASSTHROUGHID": "passthrough_id",
-    "PASSTHROUGHNAME": "passthrough_name",
-    "PERIODCOVERED": "period_covered",
-    "PREVIOUS_COMPLETED_ON": "previous_completed_on",
-    "PREVIOUSDATEFIREWALL": "previous_date_firewall",
-    "PRIORFINDINGREFNUMS": "prior_finding_ref_nums",
-    "PROGRAMTOTAL": "program_total",
-    "PYSCHEDULE": "prior_year_schedule",
-    "QCOSTS": "questioned_costs",
-    # not in key at all double check this (cfda table)
-    "QCOSTS2": "questioned_costs",
-    # only in the key
-    "R&D": "research_and_development",
-    # only in the data columns
-    "RD": "research_and_development",
-    "REPEATFINDING": "repeat_finding",
-    # "REPORTABLECONDITION/SIGNIFICANTDEFICIENCY": "condition_or_deficiency",
-    # these are together in the key buy separate in the fy 22 data probably want to break them out in the models
-    "REPORTABLECONDITION": "condition_or_deficiency",
-    "REPORTABLECONDITION_MP": "condition_or_deficiency_major_program",
-    "SIGNIFICANTDEFICIENCY_MP": "condition_or_deficiency_major_program",
-    "REPORTABLECONDITION/SIGNIFICANTDEFICIENCY_MP": "condition_or_deficiency_major_program",
-    "REPORTID": "report_id",
-    "REPORTREQUIRED": "report_required",
-    "SEQNUM": "seqnum",
-    "SEQ_NUMBER": "seq_number",
-    "SIGNIFICANTDEFICIENCY": "significant_deficiency",
-    "SP_FRAMEWORK": "sp_framework",
-    "SP_FRAMEWORK_REQUIRED": "sp_framework_required",
-    "STATE": "state",
-    "STATECLUSTERNAME": "state_cluster_name",
-    "STREET1": "street1",
-    "STREET2": "street2",
-    "TEXT": "text",
-    "TITLE": "title",
-    "TOTFEDEXPEND": "total_fed_expenditures",
-    "TYPEOFENTITY": "type_of_entity",
-    "TYPEREPORT_FS": "type_report_financial_statements",
-    "TYPEREPORT_MP": "type_report_major_program",
-    "TYPEREPORT_SP_FRAMEWORK": "type_report_special_purpose_framework",
-    "TYPEREQUIREMENT": "type_requirement",
-    "TYPE_ID": "type_id",
-    "UEI": "uei",
-    "UEISEQNUM": "uei_seq_num",
-    "VERSION": "version",
-    "ZIPCODE": "zip_code",
-}
-
-table_mappings = {
-    "gen": "General",
-    "cfda": "CfdaInfo",
-    "findings": "Findings",
-    "findingstext_formatted": "Findingstext",
-    "captext_formatted": "Captext",
-    "notes": "Notes",
-    "cpas": "MultipleCpasInfo",
-    "revisions": "Revisions",
-    "ueis": "UeiInfo",
-    "agency": "Agencies",
-    "passthrough": "Passthrough",
-    "eins": "EinInfo",
-    "duns": "DunsInfo",
-}
-
-table_mappings_from_key = {
-    "GENERAL": "General",
-    "CFDA INFO": "CfdaInfo",
-    "FINDINGS": "Findings",
-    "FINDINGSTEXT": "Findingstext",
-    "CAPTEXT": "Captext",
-    "NOTES": "Notes",
-    "MULTIPLE CPAS INFO": "MultipleCpasInfo",
-    "REVISIONS": "Revisions",
-    "UEI INFO": "UeiInfo",
-    "AGENCIES": "Agencies",
-    "PASSTHROUGH": "Passthrough",
-    "EIN INFO": "EinInfo",
-    "DUN INFO": "DunsInfo",
-}
-
-# the doc strings have the descriptive table names from the docs, this decodes to get the table file name
-file_to_table_name_mapping = {
-    "GENERAL": "gen",
-    "CFDAINFO": "cfda",
-    "FINDINGS": "findings",
-    "FINDINGSTEXT": "findingstext_formatted",
-    "CAPTEXT": "captext_formatted",
-    "NOTES": "notes",
-    "MULTIPLECPASINFO": "cpas",
-    "REVISIONS": "revisions",
-    "UEIINFO": "ueis",
-    "AGENCIES": "agency",
-    "PASSTHROUGH": "passthrough",
-    "EININFO": "eins",
-    "DUNINFO": "duns",
-}
-
-boolen_fields = [
-    "arra",
-    "direct",
-    "loans",
-    "major_program",
-    "passthrough_award",
-    "research_and_development",
-    "material_weakness",
-    "modified_opinion",
-    "other_findings",
-    "other_non_compliance",
-    "questioned_cost",
-    "repeat_finding",
-    "significant_deficiency",
-    "condition_or_deficiency",
-    "condition_or_deficiency_major_program",
-    "current_or_former_findings",
-    "dup_reports",
-    "going_concern",
-    "low_risk",
-    "material_noncompliance",
-    "material_weakness",
-    "material_weakness_major_program",
-    "multiple_cpas",
-    "multiple_duns",
-    "multiple_eins",
-    "multiple_ueis",
-    "prior_year_schedule",
-    "questioned_costs",
-    "reportable_condition",
-    "report_required",
-    "sp_framework_required",
-    "charts_tables",
-]
