+++ backend/data_distro/mappings/upload_dictonaries.py
@@ -1,0 +1,69 @@
+# List of the tables to ingest
+tables = [
+    "gen",
+    "cfda",
+    "findings",
+    "findingstext_formatted",
+    "captext_formatted",
+    "notes",
+    "cpas",
+    "revisions",
+    "ueis",
+    "agency",
+    "passthrough",
+    "eins",
+    "duns",
+]
+
+# The doc strings have the descriptive table names from the docs, this decodes to get the table file name
+file_to_table_name_mapping = {
+    "GENERAL": "gen",
+    "CFDAINFO": "cfda",
+    "FINDINGS": "findings",
+    "FINDINGSTEXT": "findingstext_formatted",
+    "CAPTEXT": "captext_formatted",
+    "NOTES": "notes",
+    "MULTIPLECPASINFO": "cpas",
+    "REVISIONS": "revisions",
+    "UEIINFO": "ueis",
+    "AGENCIES": "agency",
+    "PASSTHROUGH": "passthrough",
+    "EININFO": "eins",
+    "DUNINFO": "duns",
+}
+
+# Boolean fields
+boolen_fields = [
+    "arra",
+    "direct",
+    "loans",
+    "major_program",
+    "passthrough_award",
+    "research_and_development",
+    "material_weakness",
+    "modified_opinion",
+    "other_findings",
+    "other_non_compliance",
+    "questioned_cost",
+    "repeat_finding",
+    "significant_deficiency",
+    "condition_or_deficiency",
+    "condition_or_deficiency_major_program",
+    "current_or_former_findings",
+    "dup_reports",
+    "going_concern",
+    "low_risk",
+    "material_noncompliance",
+    "material_weakness",
+    "material_weakness_major_program",
+    "multiple_cpas",
+    "multiple_duns",
+    "multiple_eins",
+    "multiple_ueis",
+    "prior_year_schedule",
+    "questioned_costs",
+    "reportable_condition",
+    "report_required",
+    "sp_framework_required",
+    "charts_tables",
+]
