+++ backend/data_distro/mappings/upload_mapping.py
@@ -1,164 +1,8 @@
-# inspect the models to create a script that crates a key we can use to upload the data
-from django.apps import apps
-
-from data_distro.download_model_dictonaries import (
-    table_mappings,
-    file_to_table_name_mapping,
-)
-
-
-# From a given table, look at the column name and know where to load the data
-sample_upload_mapping_structure = {
-    "table file name": {
-        "column_name": ["model_name", "django_field_name"],
-        "column_name2": ["model_name", "django_field_name2"],
-    }
-}
-
-
-def make_table_structure():
-    add_realtional = []
-    blank_help = []
-    leftovers = []
-    # preload tables into a dict
-    upload_mapping = {}
-    for table_title in table_mappings.keys():
-        upload_mapping[table_title] = {}
-
-    new_fields = [
-        # django generates
-        "id",
-        # I added
-        "is_public",
-        "data_source",
-        "create_date",
-        "modified_date",
-        # relational links these will change if you move fields around
-        "general",
-        "findings_text",
-    ]
-
-    leftovers = []
-
-    distro_classes = apps.all_models["data_distro"]
-    # this should be enough to make a key
-    for model in distro_classes:
-        mod_class = distro_classes[model]
-        mod_name = mod_class.__name__
-        fields = mod_class._meta.get_fields()
-        for field in fields:
-            f_name = field.name
-            try:
-                help_text = field.help_text
-            except AttributeError:
-                help_text = ""
-                add_realtional.append([f_name, model])
-                new_fields.append(f_name)
-            if help_text != "":
-                help_text = field.help_text
-                source = help_text.split("Census mapping: ", 1)[1]
-                table_doc_name = source.split(", ", 1)[0]
-                column_name = source.split(", ", 1)[1]
-                table_file_name = file_to_table_name_mapping[
-                    table_doc_name.upper().replace(" ", "")
-                ]
-                upload_mapping[table_file_name][column_name] = [mod_name, f_name]
-            else:
-                if f_name not in new_fields:
-                    blank_help.append(f_name)
-                else:
-                    # just a check
-                    leftovers.append(f_name)
-
-    if len(blank_help) > 0:
-        print("~ WARNING ~ Check blank fields: blank_help={0}".format(blank_help))
-
-    # this should relational fields
-    print(
-        "Fields with no help (relational and array models that need some custom logic for loading): add_realtional={0}".format(
-            add_realtional
-        )
-    )
-
-    return upload_mapping, add_realtional
-
-
-# last run's results
-add_realtional = [
-    ["general", "auditee"],
-    ["general", "auditor"],
-    ["general", "cfdainfo"],
-    ["findings", "findingstext"],
-    ["general", "findings"],
-    ["general", "captext"],
-    ["general", "notes"],
-    ["general", "revisions"],
-    ["general", "agencies"],
-    ["general", "passthrough"],
-]
-
-# these should be relational fields and list fields
-blank_help = [
-    "duns_list",
-    "uei_list",
-    "auditor",
-    "agencies",
-    "auditee",
-    "cfda",
-    "cap_text",
-    "notes",
-    "revisions",
-    "passthrough",
-    "auditor",
-    "agency",
-]
-
-# just to confirm all data is accounted for as expected
-leftovers = [
-    "general",
-    "id",
-    "is_public",
-    "general",
-    "id",
-    "is_public",
-    "general",
-    "id",
-    "is_public",
-    "findings",
-    "id",
-    "is_public",
-    "general",
-    "id",
-    "findings_text",
-    "is_public",
-    "general",
-    "id",
-    "is_public",
-    "general",
-    "id",
-    "is_public",
-    "general",
-    "id",
-    "is_public",
-    "general",
-    "id",
-    "is_public",
-    "general",
-    "id",
-    "is_public",
-    "id",
-    "general",
-    "id",
-    "general",
-    "id",
-    "findings",
-    "is_public",
-    "modified_date",
-    "create_date",
-    "data_source",
-]
-
-# mapping for upload
+"""
+Mapping for upload.
+Used by public_data_loader.
+Update with create_upload_mapping management command.
+"""
 upload_mapping = {
     "gen": {
         "AUDITEECERTIFYNAME": ["Auditee", "auditee_certify_name"],
