+++ backend/manifests/task-manifest-staging.yaml
@@ -1,0 +1,23 @@
+---
+applications:
+- name: task-runner
+  buildpacks:
+    - python_buildpack
+  memory: 2G
+  path: ../
+  timeout: 180
+  health-check-type: process
+  env:
+    ENV: STAGING
+    DJANGO_BASE_URL: https://fac-staging.app.cloud.gov
+  services:
+  - fac-db
+  - fac-public-s3
+  - fac-key-service
+  command: (python manage.py public_data_loader -f findings.txt && echo SUCCESS || echo TASK_FAIL) && sleep infinity
+
+  # run with
+  # cf push -f manifests/task-manifest-staging.yaml
+  #
+  # This is an example task, you can modify the command to run other tasks
+
